+++ src/Exceptions/TelegramDeepLinkException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace AlexItDev91\LaravelTelegramBot\Exceptions;
+
+class TelegramDeepLinkException extends TelegramBotException
+{
+    //
+}